+++ cognite/neat/_rules/transformers/_converters.py
@@ -243,7 +243,7 @@ class InformationToDMS(ConversionTransformer[InformationRules, DMSRules]):
     """Converts InformationRules to DMSRules."""
 
     def __init__(
-        self, ignore_undefined_value_types: bool = False, reserved_properties: Literal["error", "skip"] = "error"
+        self, ignore_undefined_value_types: bool = False, reserved_properties: Literal["error", "warning"] = "error"
     ):
         self.ignore_undefined_value_types = ignore_undefined_value_types
         self.reserved_properties = reserved_properties
@@ -509,6 +509,8 @@ class ToSolutionModel(ToExtensionModel):
                     bool(prop.view.external_id.startswith(self.view_prefix)),
                 )
             )
+        else:
+            new_properties.sort(key=lambda prop: (prop.view.external_id, prop.view_property))
 
         metadata = reference_rules.metadata.model_copy(
             deep=True,
@@ -1041,7 +1043,7 @@ class _InformationRulesConverter:
         self.property_count_by_container: dict[ContainerEntity, int] = defaultdict(int)
 
     def as_dms_rules(
-        self, ignore_undefined_value_types: bool = False, reserved_properties: Literal["error", "skip"] = "error"
+        self, ignore_undefined_value_types: bool = False, reserved_properties: Literal["error", "warning"] = "error"
     ) -> "DMSRules":
         from cognite.neat._rules.models.dms._rules import (
             DMSContainer,