+++ tests/tests_integration/test_session/test_graph_flow.py
@@ -39,7 +39,7 @@ class TestExtractToLoadFlow:
     def test_classic_to_dms(self, cognite_client: CogniteClient, data_regression: DataRegressionFixture) -> None:
         neat = NeatSession(cognite_client, storage="oxigraph")
         neat.read.cdf.classic.graph("Utsira", identifier="id")
-        neat.convert("dms")
+        neat.convert()
         neat.mapping.data_model.classic_to_core("Classic")
         neat.set.data_model_id(("sp_windfarm", "WindFarm", "v1"))
 
@@ -70,14 +70,14 @@ class TestExtractToLoadFlow:
     def test_dexpi_to_dms(self, cognite_client: CogniteClient, data_regression: DataRegressionFixture) -> None:
         neat = NeatSession(cognite_client)
         neat.read.xml.dexpi(DATA_FOLDER / "depxi_example.xml")
-        neat.infer(max_number_of_instance=-1)
+        neat.infer()
 
         # Hack to ensure deterministic output
         rules = neat._state.rule_store.last_verified_information_rules
         rules.metadata.created = datetime.datetime.fromisoformat("2024-09-19T00:00:00Z")
         rules.metadata.updated = datetime.datetime.fromisoformat("2024-09-19T00:00:00Z")
 
-        neat.convert("dms")
+        neat.convert()
         neat.set.data_model_id(("dexpi_playground", "DEXPI", "v1.3.1"))
 
         if True:
@@ -107,14 +107,14 @@ class TestExtractToLoadFlow:
     def test_aml_to_dms(self, cognite_client: CogniteClient, data_regression: DataRegressionFixture) -> None:
         neat = NeatSession(cognite_client)
         neat.read.xml.aml(DATA_FOLDER / "aml_example.aml")
-        neat.infer(max_number_of_instance=-1)
+        neat.infer()
 
         # Hack to ensure deterministic output
         rules = neat._state.rule_store.last_verified_information_rules
         rules.metadata.created = datetime.datetime.fromisoformat("2024-09-19T00:00:00Z")
         rules.metadata.updated = datetime.datetime.fromisoformat("2024-09-19T00:00:00Z")
 
-        neat.convert("dms")
+        neat.convert()
         neat.set.data_model_id(("aml_playground", "AML", "terminology_3.0"))
 
         if True:
@@ -169,7 +169,7 @@ class TestExtractToLoadFlow:
     def test_classic_to_cdf(self, cognite_client: CogniteClient) -> None:
         neat = NeatSession(cognite_client, storage="oxigraph")
         neat.read.cdf.classic.graph("Utsira", identifier="externalId")
-        neat.convert("dms")
+        neat.convert()
 
         neat.mapping.data_model.classic_to_core("NeatInc")
         neat.set.data_model_id(("sp_windfarm", "WindFarm", "v1"))